SearchEngineGENAI/requirements.txt
langchain
python-dotenv
ipykernel
langchain-google-genai
langchain_community
langchain_core
langchain_groq
langchain_huggingface
streamlit
google-generativeai
pymupdf
faiss-cpu
langchain_chroma
chromadb
duckdb
pandas
duckduckgo-search
arxiv
wikipedia
sentence_transformers